Move placeholder impl back into DS module and keep the core extracted
