+++ app/design-system/design-system-hedvig/build.gradle.kts
@@ -11,6 +11,7 @@ dependencies {
   api(libs.androidx.compose.foundation)
   api(libs.coil.coil)
   api(projects.designSystemApi)
+  api(projects.placeholder)
 
   implementation(libs.androidx.activity.compose)
   implementation(libs.androidx.compose.foundationLayout)
