+++ app/feature/feature-profile/src/main/kotlin/com/hedvig/android/feature/profile/tab/ProfileDestination.kt
@@ -50,6 +50,7 @@ import androidx.lifecycle.compose.dropUnlessResumed
 import com.google.accompanist.permissions.isGranted
 import com.hedvig.android.compose.ui.preview.PreviewContentWithProvidedParametersAnimatedOnClick
 import com.hedvig.android.design.system.hedvig.HedvigAlertDialog
+import com.hedvig.android.design.system.hedvig.HedvigPreview
 import com.hedvig.android.design.system.hedvig.HedvigRedTextButton
 import com.hedvig.android.design.system.hedvig.HedvigText
 import com.hedvig.android.design.system.hedvig.HedvigTheme
@@ -62,8 +63,8 @@ import com.hedvig.android.design.system.hedvig.icon.InfoFilled
 import com.hedvig.android.design.system.hedvig.icon.InfoOutline
 import com.hedvig.android.design.system.hedvig.icon.MultipleDocuments
 import com.hedvig.android.design.system.hedvig.icon.Settings
-import com.hedvig.android.design.system.hedvig.placeholder.PlaceholderHighlight
 import com.hedvig.android.design.system.hedvig.placeholder.hedvigPlaceholder
+import com.hedvig.android.placeholder.PlaceholderHighlight
 import com.hedvig.android.design.system.hedvig.placeholder.shimmer
 import com.hedvig.android.design.system.hedvig.plus
 import com.hedvig.android.memberreminders.ui.MemberReminderCards
@@ -356,6 +357,7 @@ private fun ProfileRow(
         .size(24.dp)
         .hedvigPlaceholder(
           isLoading,
+          shape = HedvigTheme.shapes.cornerSmall,
           highlight = PlaceholderHighlight.shimmer(),
         ),
     )
@@ -365,12 +367,14 @@ private fun ProfileRow(
       modifier = Modifier
         .hedvigPlaceholder(
           isLoading,
+          shape = HedvigTheme.shapes.cornerSmall,
           highlight = PlaceholderHighlight.shimmer(),
         ),
     )
   }
 }
 
+@HedvigPreview
 @Composable
 private fun PreviewProfileItemRows() {
   HedvigTheme {
