+++ app/shared/forever-ui/src/main/kotlin/com/hedvig/android/shared/foreverui/ui/ui/ForeverDestination.kt
@@ -64,8 +64,8 @@ import com.hedvig.android.design.system.hedvig.Surface
 import com.hedvig.android.design.system.hedvig.icon.Copy
 import com.hedvig.android.design.system.hedvig.icon.HedvigIcons
 import com.hedvig.android.design.system.hedvig.icon.InfoOutline
-import com.hedvig.android.design.system.hedvig.placeholder.PlaceholderHighlight
 import com.hedvig.android.design.system.hedvig.placeholder.hedvigPlaceholder
+import com.hedvig.android.placeholder.PlaceholderHighlight
 import com.hedvig.android.design.system.hedvig.placeholder.shimmer
 import com.hedvig.android.design.system.hedvig.rememberHedvigBottomSheetState
 import com.hedvig.android.language.LanguageService
@@ -220,6 +220,7 @@ internal fun LoadingForeverContent() {
         .size(215.dp)
         .hedvigPlaceholder(
           visible = true,
+          shape = CircleShape,
           highlight = PlaceholderHighlight.shimmer(),
         ),
     )
