+++ app/ui/placeholder/src/main/AndroidManifest.xml
