+++ app/ui/ui-emergency/build.gradle.kts
@@ -18,5 +18,4 @@ dependencies {
   implementation(projects.coreUiData)
   implementation(projects.dataContractAndroid)
   implementation(projects.designSystemHedvig)
-  implementation(projects.placeholderHedvig)
 }
